Move ListNode to the correct namespace
